proxy: move ProxyConfig to separate file
It's moved to host config and it's validations.
